+++ src/test/php/lang/mirrors/unittest/FieldTest.class.php
@@ -6,6 +6,8 @@ use lang\mirrors\TypeMirror;
 use lang\IllegalArgumentException;
 
 class FieldTest extends AbstractFieldTest {
+
+  /** @type lang.mirrors.Field */
   private $fixture;
 
   #[@test]
@@ -41,5 +43,13 @@ class FieldTest extends AbstractFieldTest {
   #[@test]
   public function type_fields_declaring_type() {
     $this->assertEquals($this->type->parent(), $this->fixture('type')->declaredIn());
+  }
+
+  #[@test]
+  public function string_representation() {
+    $this->assertEquals(
+      'lang.mirrors.Field(private lang.mirrors.Field $fixture)',
+      $this->fixture('fixture')->toString()
+    );
   }
 }