+++ ChangeLog.md
@@ -3,8 +3,9 @@ Mirrors change log
 
 ## ?.?.? / ????-??-??
 
-* Changed lang.mirrors.Field::get() to read() - @thekid
-* Changed lang.mirrors.Field::set() to modify() - @thekid
+* Implemented `toString()` and string casts for Field class - @thekid
+* Changed lang.mirrors.Field::get() to `read()` - @thekid
+* Changed lang.mirrors.Field::set() to `modify()` - @thekid
 
 ## 0.1.1 / 2015-04-12